+++ app/assets/javascripts/folio/input/tiptap.js
@@ -21,6 +21,7 @@ window.Folio.Stimulus.register('f-input-tiptap', class extends window.Stimulus.C
     newRecord: { type: Boolean, default: false },
     placementType: String,
     placementId: Number,
+    attributeName: String,
     latestRevisionAt: String,
     hasUnsavedChanges: { type: Boolean, default: false }
   }
@@ -149,7 +150,7 @@ window.Folio.Stimulus.register('f-input-tiptap', class extends window.Stimulus.C
       this.inputTarget.value = ''
     }
 
-    this.dispatch('updateWordCount', { detail: { wordCount } })
+    this.dispatch('updateWordCount', { detail: { wordCount, attributeName: this.attributeNameValue } })
 
     if (!this.ignoreValueChangesValue) {
       this.inputTarget.dispatchEvent(new window.Event('change', { bubbles: true }))
@@ -196,7 +197,7 @@ window.Folio.Stimulus.register('f-input-tiptap', class extends window.Stimulus.C
         words: value[valueKeys.word_count],
         characters: value[valueKeys.character_count]
       })
-      this.dispatch('updateWordCount', { detail: { wordCount } })
+      this.dispatch('updateWordCount', { detail: { wordCount, attributeName: this.attributeNameValue } })
     }
 
     const data = {
@@ -372,7 +373,8 @@ window.Folio.Stimulus.register('f-input-tiptap', class extends window.Stimulus.C
 
     const data = {
       tiptap_revision: {
-        content: this.latestContent
+        content: this.latestContent,
+        attribute_name: this.attributeNameValue || 'tiptap_content'
       },
       placement: {
         type: this.placementTypeValue,